Implement kernel command-line search and replacement
Implement kernel command-line search/replace functionality.  This is for
the case of needing to replace some bootloader arguments, but not others.
This can be used if a bootloader is known to misreport the hardware due
to glitches or poor bootloader behavior.

Change-Id: Ic50b1f0147eaa4c3cf54424b6322494adf96e9de
(cherry picked from commit 2702b2ecae80dcfe20248484c30d16e37937d698)